move date format function
